+++ packages/sling-helpers/src/sdk/onSdkAccessible.js
@@ -1,3 +1,4 @@
+import '@babel/polyfill';
 import * as domHelper from '../dom/domHelper.js';
 
 export const onSdkAccessible = async () => {