+++ packages/sling-web-component-brand-icon/src/index.js
@@ -1,0 +1,4 @@
+import { domHelper } from 'sling-helpers';
+import { BrandIcon } from './component/BrandIcon.js';
+
+domHelper.registerComponent('sling-brand-icon', BrandIcon);