+++ packages/sling-web-component-sdk-connect/src/index.js
@@ -1,0 +1,4 @@
+import { domHelper } from 'sling-helpers';
+import { SdkConnect } from './component/SdkConnect.js';
+
+domHelper.registerComponent('sling-sdk-connect', SdkConnect);