+++ packages/sling-web-component-snackbar/index.js
@@ -1,0 +1,2 @@
+import 'sling-web-component-button';
+import './src/index.js';